refactor(test) : 테스트 파일 패키지 단위로 재배치
